matrix and measurer
