+++ packages/perseus/src/widget-ai-utils/matrix/matrix-ai-utils.test.ts
@@ -3,6 +3,8 @@ import {userEvent as userEventLib} from "@testing-library/user-event";
 
 import {renderQuestion} from "../../widgets/__testutils__/renderQuestion";
 
+import {getPromptJSON} from "./matrix-ai-utils";
+
 import type {PerseusRenderer} from "../../perseus-types";
 import type {UserEvent} from "@testing-library/user-event";
 
@@ -33,11 +35,42 @@ const question: PerseusRenderer = {
     },
 };
 
-describe("matrix widget", () => {
+describe("Matrix AI utils", () => {
     let userEvent: UserEvent;
     beforeEach(() => {
         userEvent = userEventLib.setup({
             advanceTimers: jest.advanceTimersByTime,
+        });
+    });
+
+    it("it returns JSON with the expected format and fields", () => {
+        const renderProps: any = {
+            matrixBoardSize: [4, 3],
+        };
+
+        const userInput: any = {
+            answers: [
+                [1, 2, 3, 4],
+                [5, 6, 7, 8],
+                [9, 10, 11, 12],
+            ],
+        };
+
+        const resultJSON = getPromptJSON(renderProps, userInput);
+
+        expect(resultJSON).toEqual({
+            type: "matrix",
+            options: {
+                width: 3,
+                height: 4,
+            },
+            userInput: {
+                answerRows: [
+                    [1, 2, 3, 4],
+                    [5, 6, 7, 8],
+                    [9, 10, 11, 12],
+                ],
+            },
         });
     });
